scripts/requirements.txt
feedparser
requests
python-dateutil
beautifulsoup4
tzdata
google-genai